シナリオ等調整。カメラ系以外完成

+++ sdk/workspace/BeachFlags/detection/detectAngleForCurrentTargetValue.h
@@ -14,7 +14,7 @@ public:
 
 private:
 	CalculateAngle mCalculateAngle;
-	double mTargetAngle; // 単位：度
+	int mTargetAngle; // 単位：度
 };
 
 #endif
